using MNIST database
使用tanh为激活函数

+++ main.h
@@ -1,3 +1,4 @@
 #pragma once
-using valT = double;
+// using valT = double;
+using valT = long double;
 using funcT = valT (*)(valT);
